Reference

Contents

Reference

Contents

Index

Index

MGA.AlternativeSolutions

MGA.add_solution!

MGA.create_alternative_generating_problem!

MGA.extract_variables

MGA.generate_alternatives!

MGA.update_solutions!

MGA.AlternativeSolutions — Type

Structure holding the solutions for the near-optimal alternatives.

MGA.add_solution! — Method

add_solution!(
    model::JuMP.Model,
    metric::Distances.Metric
)

Add a previously found solution to a modelling-for-generating-alternatives problem. Used for iteratively finding multiple alternative solutions.

MGA.create_alternative_generating_problem! — Method

create_alternative_problem(
    model::JuMP.Model,
    optimality_gap::Float64,
    metric::Distances.Metric,
    [selected_variables::Vector{VariableRef}]
)

Transform a JuMP model into a model solving its corresponding modelling-for-generating-alternatives problem.

Arguments

model::JuMP.Model: a solved JuMP model for which alternatives are generated.

optimality_gap::Float64: the maximum percentage deviation (>= 0) an alternative may have compared to the optimal solution.

metric::Distances.Metric=SqEuclidean(): the metric used to maximise the difference between alternatives and the optimal solution.

fixed_variables::Vector{VariableRef}=[]: a subset of all variables of model that are not allowed to be changed when seeking for alternatives.

MGA.extract_variables — Method

extract_variables(model::JuMP.Model)

Retrieve the variables and their solution values from a JuMP model.

MGA.generate_alternatives! — Method

result = generate_alternatives!(
  model::JuMP.Model,
  optimality_gap::Float64,
  n_alternatives::Int64;
  metric::Distances.Metric = SqEuclidean(),
  selected_variables::Vector{VariableRef} = []
)

Generate n_alternatives solutions to model which are as distant from the optimum and each other, but with a maximum optimality_gap.

Arguments

model::JuMP.Model: a solved JuMP model for which alternatives are generated.

optimality_gap::Float64: the maximum percentage deviation (>=0) an alternative may have compared to the optimal solution.

n_alternatives: the number of alternative solutions sought.

metric::Distances.Metric=SqEuclidean(): the metric used to maximise the difference between alternatives and the optimal solution.

fixed_variables::Vector{VariableRef}=[]: a subset of all variables of model that are not allowed to be changed when seeking for alternatives.

MGA.update_solutions! — Method

update_solutions!(results::AlternativeSolutions, model::JuMP.Model)

Update the set of results AlternativeSolutions with the variable values obtained when solving the JuMP model model.